backend/requirements.txt
Flask==2.2.2
Werkzeug>=2.2.2,<2.3
Flask-Cors==3.0.10
sqlalchemy==1.4.46
pymysql==1.0.2
gunicorn
python-dotenv
